style: rename testimonials
