--- controller/testimonialController.js
@@ -1,10 +1,0 @@
-const service = require('../services/testimonial');
-
-const getAllTestimonials = async () => {
-  const response = await service.getAllTestimonials();
-  return response;
-};
-
-module.exports = {
-  getAllTestimonials,
-};
