--- services/testimonial.js
@@ -1,7 +1,0 @@
-const testimonials = require('../mock/testimonials-JSON');
-
-const getAllTestimonials = async () => testimonials;
-
-module.exports = {
-  getAllTestimonials,
-};
