+++ services/testimonies.js
@@ -1,0 +1,5 @@
+const getAllTestimonies = async () => [];
+
+module.exports = {
+  getAllTestimonies,
+};
